+++ test/test.js
@@ -47,6 +47,14 @@ describe('finalhandler(req, res)', function () {
       .expect(404, '', done)
     })
 
+    it('should include security header', function (done) {
+      var server = createServer()
+      request(server)
+      .get('/foo')
+      .expect('X-Content-Type-Options', 'nosniff')
+      .expect(404, done)
+    })
+
     it('should not hang/error if there is a request body', function (done) {
       var buf = new Buffer(1024 * 16)
       var server = createServer()
@@ -60,7 +68,7 @@ describe('finalhandler(req, res)', function () {
   })
 
   describe('error response', function () {
-    it('include error stack', function (done) {
+    it('should include error stack', function (done) {
       var server = createServer(new Error('boom!'))
       request(server)
       .get('/foo')
@@ -72,6 +80,14 @@ describe('finalhandler(req, res)', function () {
       request(server)
       .head('/foo')
       .expect(404, '', done)
+    })
+
+    it('should include security header', function (done) {
+      var server = createServer(new Error('boom!'))
+      request(server)
+      .get('/foo')
+      .expect('X-Content-Type-Options', 'nosniff')
+      .expect(500, done)
     })
 
     it('should handle non-error-objects', function (done) {